test policy in delete debt method

+++ app/Http/Controllers/DebtController.php
@@ -154,14 +154,18 @@ class DebtController extends Controller
      */
     public function destroy(Request $request, Debt $debt, ShareService $shareService): RedirectResponse
     {
-        $validated = Validator::make($request->all(), [
-            'id' => ['required', 'numeric', 'exists:debts,id', new IsDebtOwner],
-        ])->validate();
+        if ($request->user()->cannot('delete', $debt)) {
+            return redirect()->route('debt.index')->withErrors(['id' => "You do not have permission to delete this debt."]);
+        } else {
+            $validated = Validator::make($request->all(), [
+                'id' => ['required', 'numeric', 'exists:debts,id'],
+            ])->validate();
 
-        $shareService->deleteDebtShares($debt);
+            $shareService->deleteDebtShares($debt);
 
-        $debt->delete();
+            $debt->delete();
 
-        return redirect()->route('debt.index')->with('status', "Debt deleted successfully.");;
+            return redirect()->route('debt.index')->with('status', "Debt deleted successfully.");
+        }
     }
 }
